Splited HDL ATmegaxx0_1 GPIO Defs into several files.
