+++ hdl/avr/mega/xx0_1/include/reg.hpp
@@ -1,322 +1,31 @@
 #ifndef HDL_AVR_MEGA_XX0_1_REG_HPP
 #define HDL_AVR_MEGA_XX0_1_REG_HPP
 
+#include <util/memory_mapping.hpp>
+#include <util/dummy.hpp>
+#include "../../../common/include/reg.hpp"
+
+
 #if defined(__AVR_ATmega640__) || defined(__AVR_ATmega1280__) || defined(__AVR_ATmega2560__)
 # define __ATmegaxx0__
 #elif defined(__AVR_ATmega1281__) || defined(__AVR_ATmega2561__)
 # define __ATmegaxx1__
 #endif
 
-#include <util/memory_mapping.hpp>
+#include "gpio/fields.hpp"
+#include "gpio/regs.hpp"
+#include "gpio/modules.hpp"
+#include "gpio/defs.hpp"
 
 namespace HDL {
   using namespace MemoryMapping;
-  
-  namespace Private {
-    //! \brief This is a utility type for templates
-    /*! 
-     * Due to C++ current (october 2015) specifications,
-     * templates cannot be specialised in the class where 
-     * it has been declared.
-     * But partial specialisations are allowed.
-     * So this type is used when the developper  
-     * wants to specialise a template in the same class
-     * as it has been declared.
-     */ 
-    using DummyType = u8;
-
-    //! \brief This value should be used with DummyType
-    /*!
-     * It should even be the only acceptable value of DummyType.
-     */
-    constexpr u8 DUMMY_VALUE = 0;
-  }
-  
-  //! \brief Definitions that are common to all AVR MCUs
-  namespace AVR {
-    
-    namespace Private {
-      
-      //! \brief Transform an IO address into a MEM address
-      //! \param addr : IO address, can take a value between 0x0000 and 0x003F.
-      //! \return MEM address
-      static constexpr usys SFR_IO(usys addr) {
-	return 0x0020 + addr;
-      }
-      
-      //! \brief Returns a MEM address
-      //! \param addr : The address of a register in the MEM Space
-      //! \return MEM address
-      /*! 
-       * Doesn't do anything, but must be used by HDL developpers
-       * to specify that an address is a memory space address
-       * (instead of an IO space address).
-       */
-      static constexpr usys SFR_MEM(usys addr) {
-	return addr;
-      }
-      
-    }
-    
-  }
-  using namespace AVR;
     
   namespace ATMegaxx0_1 {
-      
-    namespace Private {
-      using namespace HDL::Private;
-      using namespace HDL::AVR::Private;
-
-      namespace SimpleDefs {
-  
-	struct GPIO_Undef_ModuleDefs {
-	  static constexpr bool EXIST = false;
-	};
     
-	struct GPIO_Common_ModuleDefs {
-	  static constexpr bool EXIST = true;
-    
-	  struct Groups {
-	    struct PIN;
-	    struct DDR;
-	    struct PORT;
-	  };
-
-	  static constexpr u8 PIN_MASK = 0b11111111;
-    
-	  static constexpr u8 PIN_0 = 0;
-	  static constexpr u8 PIN_1 = 1;
-	  static constexpr u8 PIN_2 = 2;
-	  static constexpr u8 PIN_3 = 3;
-	  static constexpr u8 PIN_4 = 4;
-	  static constexpr u8 PIN_5 = 5;
-	  static constexpr u8 PIN_6 = 6;
-	  static constexpr u8 PIN_7 = 7;
-
-	  static constexpr u8 DDR_0 = 0;
-	  static constexpr u8 DDR_1 = 1;
-	  static constexpr u8 DDR_2 = 2;
-	  static constexpr u8 DDR_3 = 3;
-	  static constexpr u8 DDR_4 = 4;
-	  static constexpr u8 DDR_5 = 5;
-	  static constexpr u8 DDR_6 = 6;
-	  static constexpr u8 DDR_7 = 7;
-
-	  static constexpr u8 PORT_0 = 0;
-	  static constexpr u8 PORT_1 = 1;
-	  static constexpr u8 PORT_2 = 2;
-	  static constexpr u8 PORT_3 = 3;
-	  static constexpr u8 PORT_4 = 4;
-	  static constexpr u8 PORT_5 = 5;
-	  static constexpr u8 PORT_6 = 6;
-	  static constexpr u8 PORT_7 = 7;
-	};
-
-	struct GPIO_A_ModuleDefs : GPIO_Common_ModuleDefs {
-	  static constexpr usys PIN  = SFR_IO(0x0000);
-	  static constexpr usys DDR  = SFR_IO(0x0001);
-	  static constexpr usys PORT = SFR_IO(0x0002);
-	};
-
-	struct GPIO_B_ModuleDefs : GPIO_Common_ModuleDefs {
-	  static constexpr usys PIN  = SFR_IO(0x0003);
-	  static constexpr usys DDR  = SFR_IO(0x0004);
-	  static constexpr usys PORT = SFR_IO(0x0005);
-	};
-
-	struct GPIO_C_ModuleDefs : GPIO_Common_ModuleDefs {
-	  static constexpr usys PIN  = SFR_IO(0x0006);
-	  static constexpr usys DDR  = SFR_IO(0x0007);
-	  static constexpr usys PORT = SFR_IO(0x0008);
-	};
-
-	struct GPIO_D_ModuleDefs : GPIO_Common_ModuleDefs {
-	  static constexpr usys PIN  = SFR_IO(0x0009);
-	  static constexpr usys DDR  = SFR_IO(0x000A);
-	  static constexpr usys PORT = SFR_IO(0x000B);
-	};
-
-	struct GPIO_E_ModuleDefs : GPIO_Common_ModuleDefs {
-	  static constexpr usys PIN  = SFR_IO(0x000C);
-	  static constexpr usys DDR  = SFR_IO(0x000D);
-	  static constexpr usys PORT = SFR_IO(0x000E);
-	};
-
-	struct GPIO_F_ModuleDefs : GPIO_Common_ModuleDefs {
-	  static constexpr usys PIN  = SFR_IO(0x000F);
-	  static constexpr usys DDR  = SFR_IO(0x0010);
-	  static constexpr usys PORT = SFR_IO(0x0011);
-	};
-
-	// warning : This GPIO only contains 6 pins
-	struct GPIO_G_ModuleDefs : GPIO_Common_ModuleDefs {
-	  struct Groups {
-	    struct PIN;
-	    struct DDR;
-	    struct PORT;
-	  };
-
-	  static constexpr u8 PIN_MASK = 0b00111111;
-    
-	  static constexpr usys PIN  = SFR_IO(0x0012);
-	  static constexpr usys DDR  = SFR_IO(0x0013);
-	  static constexpr usys PORT = SFR_IO(0x0014);
-	};
-
-#if defined(__ATmegaxx0__)
-  
-	struct GPIO_H_ModuleDefs : GPIO_Common_ModuleDefs {
-	  static constexpr usys PIN  = SFR_MEM(0x0100);
-	  static constexpr usys DDR  = SFR_MEM(0x0101);
-	  static constexpr usys PORT = SFR_MEM(0x0102);
-	};
-
-	struct GPIO_I_ModuleDefs : GPIO_Undef_ModuleDefs {};
-    
-	struct GPIO_J_ModuleDefs : GPIO_Common_ModuleDefs {
-	  static constexpr usys PIN  = SFR_MEM(0x0103);
-	  static constexpr usys DDR  = SFR_MEM(0x0104);
-	  static constexpr usys PORT = SFR_MEM(0x0105);
-	};
-
-	struct GPIO_K_ModuleDefs : GPIO_Common_ModuleDefs {
-	  static constexpr usys PIN  = SFR_MEM(0x0106);
-	  static constexpr usys DDR  = SFR_MEM(0x0107);
-	  static constexpr usys PORT = SFR_MEM(0x0108);
-	};
-
-	struct GPIO_L_ModuleDefs : GPIO_Common_ModuleDefs {
-	  static constexpr usys PIN  = SFR_MEM(0x0109);
-	  static constexpr usys DDR  = SFR_MEM(0x010A);
-	  static constexpr usys PORT = SFR_MEM(0x010B);
-	};
-
-#else
-
-	struct GPIO_H_ModuleDefs : GPIO_Undef_ModuleDefs {};
-	struct GPIO_I_ModuleDefs : GPIO_Undef_ModuleDefs {};
-	struct GPIO_J_ModuleDefs : GPIO_Undef_ModuleDefs {};
-	struct GPIO_K_ModuleDefs : GPIO_Undef_ModuleDefs {};
-	struct GPIO_L_ModuleDefs : GPIO_Undef_ModuleDefs {};
-
-#endif   //(__ATmegaxx0__)
-  
-      }
-    }
-  
     namespace Private {
       using namespace SimpleDefs;
       
       namespace GPIO {
-	template<typename ModuleDefs>
-	struct GPIO_ModuleRegisters {
-	  using Defs = ModuleDefs;
-
-	  static constexpr Register8<typename Defs::Groups::PIN>  PIN  = Defs::PIN;
-	  static constexpr Register8<typename Defs::Groups::DDR>  DDR  = Defs::DDR;
-	  static constexpr Register8<typename Defs::Groups::PORT> PORT = Defs::PORT;
-	};
-  
-	template<typename ModuleDefs>
-	struct GPIO_ModuleFields {
-	  using Defs = ModuleDefs;
-	  using Registers = GPIO_ModuleRegisters<Defs>;
-
-	  template<u8 PIN_NUM> struct Check {
-	    static_assert(Defs::PIN_MASK & (1 << PIN_NUM), "Invalid PIN_NUM");
-	  };
-    
-	  template<u8 PIN_NUM>
-	  struct PIN : Check<PIN_NUM> {
-	    static constexpr BitField8<typename Defs::Groups::PIN, PIN_NUM> field = Registers::PIN;
-	  };
-    
-	  template<u8 PIN_NUM>
-	  struct DDR : Check<PIN_NUM> {
-	    static constexpr BitField8<typename Defs::Groups::DDR, PIN_NUM> field = Registers::DDR;
-	  };
-    
-	  template<u8 PIN_NUM>
-	  struct PORT : Check<PIN_NUM> {
-	    static constexpr BitField8<typename Defs::Groups::PORT, PIN_NUM> field = Registers::PORT;
-	  };
-	};
-  
-	template<typename ModuleDefs>
-	struct GPIO_ModuleAliases {
-	  using Defs = ModuleDefs;
-	  using Fields = GPIO_ModuleFields<Defs>;
-    
-	  template<u8 MASK = Defs::PIN_MASK, DummyType DUMMY = DUMMY_VALUE> struct _PIN_Fields {};
-	  template<u8 MASK = Defs::PIN_MASK, DummyType DUMMY = DUMMY_VALUE> struct _DDR_Fields {};
-	  template<u8 MASK = Defs::PIN_MASK, DummyType DUMMY = DUMMY_VALUE> struct _PORT_Fields {};
-
-	  template<DummyType DUMMY>
-	  struct _PIN_Fields<0b00111111, DUMMY> {
-	    static constexpr auto& PIN_0 = Fields::template PIN<0>::field;
-	    static constexpr auto& PIN_1 = Fields::template PIN<1>::field;
-	    static constexpr auto& PIN_2 = Fields::template PIN<2>::field;
-	    static constexpr auto& PIN_3 = Fields::template PIN<3>::field;
-	    static constexpr auto& PIN_4 = Fields::template PIN<4>::field;
-	    static constexpr auto& PIN_5 = Fields::template PIN<5>::field;
-	  };
-
-	  template<DummyType DUMMY>
-	  struct _PIN_Fields<0b11111111, DUMMY> : _PIN_Fields<0b00111111, DUMMY> {
-	    static constexpr auto& PIN_6 = Fields::template PIN<6>::field;
-	    static constexpr auto& PIN_7 = Fields::template PIN<7>::field;
-	  };
-
-	  template<DummyType DUMMY>
-	  struct _DDR_Fields<0b00111111, DUMMY> {
-	    static constexpr auto& DDR_0 = Fields::template DDR<0>::field;
-	    static constexpr auto& DDR_1 = Fields::template DDR<1>::field;
-	    static constexpr auto& DDR_2 = Fields::template DDR<2>::field;
-	    static constexpr auto& DDR_3 = Fields::template DDR<3>::field;
-	    static constexpr auto& DDR_4 = Fields::template DDR<4>::field;
-	    static constexpr auto& DDR_5 = Fields::template DDR<5>::field;
-	  };
-
-	  template<DummyType DUMMY>
-	  struct _DDR_Fields<0b11111111, DUMMY> : _DDR_Fields<0b00111111, DUMMY> {
-	    static constexpr auto& DDR_6 = Fields::template DDR<6>::field;
-	    static constexpr auto& DDR_7 = Fields::template DDR<7>::field;
-	  };
-
-	  template<DummyType DUMMY>
-	  struct _PORT_Fields<0b00111111, DUMMY> {
-	    static constexpr auto& PORT_0 = Fields::template PORT<0>::field;
-	    static constexpr auto& PORT_1 = Fields::template PORT<1>::field;
-	    static constexpr auto& PORT_2 = Fields::template PORT<2>::field;
-	    static constexpr auto& PORT_3 = Fields::template PORT<3>::field;
-	    static constexpr auto& PORT_4 = Fields::template PORT<4>::field;
-	    static constexpr auto& PORT_5 = Fields::template PORT<5>::field;
-	  };
-
-	  template<DummyType DUMMY>
-	  struct _PORT_Fields<0b11111111, DUMMY> : _PORT_Fields<0b00111111, DUMMY> {
-	    static constexpr auto& PORT_6 = Fields::template PORT<6>::field;
-	    static constexpr auto& PORT_7 = Fields::template PORT<7>::field;
-	  };
-
-	  using PIN_Fields = _PIN_Fields<Defs::PIN_MASK>;
-	  using DDR_Fields = _DDR_Fields<Defs::PIN_MASK>;
-	  using PORT_Fields = _PORT_Fields<Defs::PIN_MASK>;
-	};
-
-	template<typename Defs>
-	struct _GPIO {
-	  static_assert(Defs::EXIST, "GPIO does not exists");
-	  struct Module
-	    : GPIO_ModuleRegisters<Defs> {
-	    struct Fields
-	      : GPIO_ModuleFields<Defs>,
-		GPIO_ModuleAliases<Defs>::PIN_Fields,
-		GPIO_ModuleAliases<Defs>::DDR_Fields,
-		GPIO_ModuleAliases<Defs>::PORT_Fields {
-	    };
-	  };
-	};
 
 	template<u8 ID> struct _GPIO_Defs : GPIO_Undef_ModuleDefs {};
 	template<> struct _GPIO_Defs<0>  : GPIO_A_ModuleDefs {};
